Rename new foldable sidebar
From Sidebar2 to Sidebar
