+++ doc/web-doc/src/components/Sidebar/menu.config.js
